server_app/requirements.txt
flask
kubernetes
pyyaml
jinja2
ultralytics